Add canBeCompleted and completeOrder functions
This adds two new methods to the OrderContract and AbstractOrder model.

It adds a method that is supposed to describe if we can run the completeOrder function.
This differs from 'isCompletable' by the fact that isCompletable does not account for ithe order is already completed or not.

This is to ensure orders are only completable if they are completable and in a not completed setting.